+++ torchvision/prototype/transforms/_type_conversion.py
@@ -56,11 +56,5 @@ class ToImagePIL(Transform):
 
 # We changed the names to align them with the new naming scheme. Still, `PILToTensor` and `ToPILImage` are
 # prevalent and well understood. Thus, we just alias them without deprecating the old names.
-
-
-class PILToTensor(ToImageTensor):
-    pass
-
-
-class ToPILImage(ToImagePIL):
-    pass
+PILToTensor = ToImageTensor
+ToPILImage = ToImagePIL
