+++ torchvision/prototype/transforms/functional/__init__.py
@@ -5,6 +5,9 @@ from ._meta import (
     convert_color_space_image_tensor,
     convert_color_space_image_pil,
     convert_color_space,
+    get_dimensions,
+    get_image_num_channels,
+    get_image_size,
 )  # usort: skip
 
 from ._augment import erase, erase_image_pil, erase_image_tensor
@@ -119,6 +122,5 @@ from ._type_conversion import (
     to_image_tensor,
     to_pil_image,
 )
-from ._utils import get_dimensions, get_image_num_channels, get_image_size
 
 from ._deprecated import rgb_to_grayscale, to_grayscale  # usort: skip
